fileRgIdx uint64->uint32

+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/metadata/domain/RowGroup.java
@@ -13,7 +13,7 @@ public class RowGroup extends Base {
     private long layoutId;
     @Getter
     @Setter
-    private long fileRgIdx;
+    private int fileRgIdx;
     @Getter
     @Setter
     private boolean isWriteBuffer;
